Use RouterTestServer in every tests that performs Network requests to make sure we don't do real requests by mistacke Hopefully might help to troubleshoot the occasionally failing test
